OpenCSG-1.3.3

+++ src/area.cpp
@@ -1,5 +1,5 @@
 // OpenCSG - library for image-based CSG rendering for OpenGL
-// Copyright (C) 2002-2011, Florian Kirsch,
+// Copyright (C) 2002-2014, Florian Kirsch,
 // Hasso-Plattner-Institute at the University of Potsdam, Germany
 //
 // This library is free software; you can redistribute it and/or 
@@ -15,7 +15,7 @@
 //
 // You should have received a copy of the GNU General Public License 
 // along with this program; if not, write to the Free Software Foundation,
-// Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
+// Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
 #include "opencsgConfig.h"
 #include "area.h"
@@ -26,13 +26,13 @@ namespace OpenCSG {
     NDCVolume::operator PCArea() const {
         using OpenGL::canvasPos;
 
-        double widthPixels  = 0.5 * (canvasPos[2] - canvasPos[0]);
-        double heightPixels = 0.5 * (canvasPos[3] - canvasPos[1]);
+        float widthPixels  = 0.5f * (canvasPos[2] - canvasPos[0]);
+        float heightPixels = 0.5f * (canvasPos[3] - canvasPos[1]);
 
-        int iminx = static_cast<int>( (minx + 1.0) * widthPixels );
-        int imaxx = static_cast<int>( (maxx + 1.0) * widthPixels );
-        int iminy = static_cast<int>( (miny + 1.0) * heightPixels );
-        int imaxy = static_cast<int>( (maxy + 1.0) * heightPixels );  
+        int iminx = static_cast<int>( (minx + 1.0f) * widthPixels );
+        int imaxx = static_cast<int>( (maxx + 1.0f) * widthPixels );
+        int iminy = static_cast<int>( (miny + 1.0f) * heightPixels );
+        int imaxy = static_cast<int>( (maxy + 1.0f) * heightPixels );  
     
         return PCArea(iminx, iminy, imaxx, imaxy); 
     }
